The French "cassation" method explained in English
This is a short extract of an unpublished article on the "cassation" method as applied by the French Cour the cassation to civil proceedings before it. The Cour de cassation is France's supreme judicial body for cases handled by 'judicial' courts, i.e., civil cases between private parties or involving the application of private law and criminal cases. For a more detailed (albeit slightly outdated) presentation of the Cour de cassation, see this page on the court's website.